+++ internal/integrationtest/bds_odh_instance_resource_test.go
@@ -33,6 +33,12 @@ var (
 	BdsInstanceOdhWithRegularComputeWorkerResourceConfig = BdsInstanceOdhResourceDependencies +
 		acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update, bdsInstanceOdhWithRegularComputeAndFlexMasterUtilRepresentation)
 
+	BdsInstanceOdhWithAddMasterUtilRequiredOnlyResource = BdsInstanceOdhResourceDependencies +
+		acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Required, acctest.Update, bdsInstanceOdhWithAddMasterUtilRepresentation)
+
+	BdsInstanceOdhWithAddMasterUtilWorkerResourceConfig = BdsInstanceOdhResourceDependencies +
+		acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update, bdsInstanceOdhWithAddMasterUtilRepresentation)
+
 	bdsInstanceOdhSingularDataSourceRepresentation = map[string]interface{}{
 		"bds_instance_id": acctest.Representation{RepType: acctest.Required, Create: `${oci_bds_bds_instance.test_bds_instance.id}`},
 	}
@@ -66,12 +72,13 @@ var (
 
 		"is_cloud_sql_configured": acctest.Representation{RepType: acctest.Optional, Create: `false`},
 		"kms_key_id":              acctest.Representation{RepType: acctest.Optional, Create: `${var.kms_key_id}`, Update: `${var.kms_key_id_for_update}`},
-		//"cloud_sql_details":       acctest.RepresentationGroup{RepType: acctest.Optional, Group: bdsInstanceNodesOdhCloudSqlRepresentation}, // capacity issue
+		//"cloud_sql_details":      acctest.RepresentationGroup{RepType: acctest.Optional, Group: bdsInstanceNodesOdhCloudSqlRepresentation}, // capacity issue
 
 		//Uncomment this when running in home region (PHX)
 		//	"defined_tags":   acctest.Representation{RepType: acctest.Optional, Create: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "value")}`, Update: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "updatedValue")}`},
 		"freeform_tags":  acctest.Representation{RepType: acctest.Optional, Create: map[string]string{"bar-key": "value"}, Update: map[string]string{"Department": "Accounting"}},
 		"network_config": acctest.RepresentationGroup{RepType: acctest.Optional, Group: bdsInstanceOdhNetworkConfigRepresentation},
+		//"os_patch_version": acctest.Representation{RepType: acctest.Optional, Update: `ol7.9-x86_64-1.24.0.100-0.0`}, // Test when patch is available
 	}
 
 	bdsInstanceOdhWithFlexComputeAndRegularMasterUtilRepresentation = acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation,
@@ -86,7 +93,6 @@ var (
 	bdsInstanceOdhWithRegularComputeAndFlexMasterUtilRepresentation = acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation,
 		map[string]interface{}{
 			"compute_only_worker_node": acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceNodesOdhUtilRepresentation}, // Regular util shape representation usable for compute worker
-			"edge_node":                acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceNodesOdhUtilRepresentation},
 		})
 
 	bdsInstanceNodesOdhCloudSqlRepresentation = map[string]interface{}{
@@ -120,13 +126,40 @@ var (
 		"number_of_nodes":          acctest.Representation{RepType: acctest.Required, Create: `2`},
 		"shape_config":             acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceNodesShapeConfigRepresentation},
 	}
+	bdsInstanceNodeFlex3ShapeRepresentation = map[string]interface{}{
+		"shape":                    acctest.Representation{RepType: acctest.Required, Create: `VM.Standard3.Flex`},
+		"subnet_id":                acctest.Representation{RepType: acctest.Required, Create: `${var.subnet_id}`},
+		"block_volume_size_in_gbs": acctest.Representation{RepType: acctest.Required, Create: `150`},
+		"number_of_nodes":          acctest.Representation{RepType: acctest.Required, Create: `2`},
+		"shape_config":             acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceNodesShapeConfigRepresentation},
+	}
 	bdsInstanceNodesShapeConfigRepresentation = map[string]interface{}{
 		"memory_in_gbs": acctest.Representation{RepType: acctest.Required, Create: `32`},
 		"ocpus":         acctest.Representation{RepType: acctest.Required, Create: `3`},
 	}
+	bdsInstanceOdhWithAddMasterUtilRepresentation = acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation,
+		map[string]interface{}{
+			// Master & Util shape should be same
+			"master_node": acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceOdhWithUpdateMasterUtilRepresentation},
+			"util_node":   acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceOdhWithUpdateMasterUtilRepresentation},
+		})
+	bdsInstanceOdhWithUpdateMasterUtilRepresentation = map[string]interface{}{
+		"shape":                    acctest.Representation{RepType: acctest.Required, Create: `VM.Standard.E4.Flex`},
+		"subnet_id":                acctest.Representation{RepType: acctest.Required, Create: `${var.subnet_id}`},
+		"block_volume_size_in_gbs": acctest.Representation{RepType: acctest.Required, Create: `150`},
+		"number_of_nodes":          acctest.Representation{RepType: acctest.Required, Update: `3`},
+		"shape_config":             acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceNodesShapeConfigRepresentation},
+	}
 	bdsInstanceOdhNetworkConfigRepresentation = map[string]interface{}{
 		"cidr_block":              acctest.Representation{RepType: acctest.Optional, Create: `111.112.0.0/16`},
 		"is_nat_gateway_required": acctest.Representation{RepType: acctest.Optional, Create: `true`},
+	}
+	bdsInstanceKafkaBrokerNodeFlexShapeRepresentation = map[string]interface{}{
+		"shape":                    acctest.Representation{RepType: acctest.Required, Create: `VM.Standard.E4.Flex`},
+		"subnet_id":                acctest.Representation{RepType: acctest.Required, Create: `${var.subnet_id}`},
+		"block_volume_size_in_gbs": acctest.Representation{RepType: acctest.Required, Create: `150`},
+		"number_of_kafka_nodes":    acctest.Representation{RepType: acctest.Required, Create: `3`},
+		"shape_config":             acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceNodesShapeConfigRepresentation},
 	}
 
 	BdsInstanceOdhResourceDependencies = acctest.GenerateResourceFromRepresentationMap("oci_core_subnet", "test_subnet", acctest.Required, acctest.Create,
@@ -180,7 +213,6 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 		acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Create, bdsInstanceOdhRepresentation), "bds", "bdsInstanceOdh", t)
 
 	acctest.ResourceTest(t, testAccCheckBdsBdsInstanceOdhDestroy, []resource.TestStep{
-
 		// verify Create with required fields
 		{
 			Config: config + compartmentIdVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr +
@@ -262,6 +294,119 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
 			),
 		},
+		// delete before next Create
+		{
+			Config: config + compartmentIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr,
+		},
+		// Create cluster with HADOOP_EXTENDED
+		{
+			Config: config + compartmentIdVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr +
+				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Create,
+					acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation, map[string]interface{}{
+						"cluster_profile": acctest.Representation{RepType: acctest.Optional, Create: `HADOOP_EXTENDED`},
+						"display_name":    acctest.Representation{RepType: acctest.Required, Create: `hadext1`, Update: `hadext1`},
+					})),
+			Check: resource.ComposeAggregateTestCheckFunc(
+				resource.TestCheckResourceAttr(resourceName, "cluster_admin_password", "T3JhY2xlVGVhbVVTQSExMjM="),
+				resource.TestCheckResourceAttrSet(resourceName, "cluster_public_key"),
+				resource.TestCheckResourceAttr(resourceName, "cluster_version", "ODH1"),
+				resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
+				resource.TestCheckResourceAttr(resourceName, "is_high_availability", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_secure", "true"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.node_type"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.shape"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.subnet_id"),
+				resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
+
+				func(s *terraform.State) (err error) {
+					resId, err = acctest.FromInstanceState(s, resourceName, "id")
+					return err
+				},
+			),
+		},
+		// Add Kafka to cluster
+		{
+			Config: config + compartmentIdVariableStr + compartmentIdUVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + kmsKeyIdUVariableStr +
+				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update,
+					acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation, map[string]interface{}{
+						"is_kafka_configured": acctest.Representation{RepType: acctest.Required, Create: `false`, Update: `true`},
+						"cluster_profile":     acctest.Representation{RepType: acctest.Optional, Create: `HADOOP_EXTENDED`, Update: `HADOOP_EXTENDED`},
+						"display_name":        acctest.Representation{RepType: acctest.Required, Create: `hadext1`, Update: `hadext1`},
+						"kafka_broker_node":   acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceKafkaBrokerNodeFlexShapeRepresentation},
+					})),
+			Check: resource.ComposeAggregateTestCheckFunc(
+				resource.TestCheckResourceAttr(resourceName, "cluster_admin_password", "T3JhY2xlVGVhbVVTQSExMjM="),
+				resource.TestCheckResourceAttrSet(resourceName, "cluster_public_key"),
+				resource.TestCheckResourceAttr(resourceName, "cluster_version", "ODH1"),
+				resource.TestCheckResourceAttr(resourceName, "is_high_availability", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_secure", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_kafka_configured", "true"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.node_type"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.shape"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.subnet_id"),
+				resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
+			),
+		},
+		// Remove Kafka to cluster
+		{
+			Config: config + compartmentIdVariableStr + compartmentIdUVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + kmsKeyIdUVariableStr +
+				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update,
+					acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation, map[string]interface{}{
+						"is_kafka_configured": acctest.Representation{RepType: acctest.Required, Create: `false`, Update: `false`},
+						"cluster_profile":     acctest.Representation{RepType: acctest.Optional, Create: `HADOOP_EXTENDED`, Update: `HADOOP_EXTENDED`},
+						"display_name":        acctest.Representation{RepType: acctest.Required, Create: `hadext1`, Update: `hadext1`},
+						"kafka_broker_node":   acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceKafkaBrokerNodeFlexShapeRepresentation},
+					})),
+			Check: resource.ComposeAggregateTestCheckFunc(
+				resource.TestCheckResourceAttr(resourceName, "cluster_admin_password", "T3JhY2xlVGVhbVVTQSExMjM="),
+				resource.TestCheckResourceAttrSet(resourceName, "cluster_public_key"),
+				resource.TestCheckResourceAttr(resourceName, "cluster_version", "ODH1"),
+				resource.TestCheckResourceAttr(resourceName, "is_high_availability", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_secure", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_kafka_configured", "false"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.node_type"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.shape"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.subnet_id"),
+				resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
+			),
+		},
+		// delete before next Create
+		{
+			Config: config + compartmentIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr,
+		},
+		// verify Create with required fields Kafka cluster
+		{
+			Config: config + compartmentIdVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies + bootstrapScriptUrlVariableStr +
+				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Required, acctest.Create,
+					acctest.RepresentationCopyWithNewProperties(bdsInstanceOdhRepresentation, map[string]interface{}{
+						"is_kafka_configured": acctest.Representation{RepType: acctest.Required, Create: `false`, Update: `false`},
+						"cluster_profile":     acctest.Representation{RepType: acctest.Optional, Create: `KAFKA`, Update: `HADOOP_EXTENDED`},
+						"display_name":        acctest.Representation{RepType: acctest.Required, Create: `kafkacluster`, Update: `kafkacluster`},
+						"kafka_broker_node":   acctest.RepresentationGroup{RepType: acctest.Required, Group: bdsInstanceKafkaBrokerNodeFlexShapeRepresentation},
+					})),
+			Check: resource.ComposeAggregateTestCheckFunc(
+				resource.TestCheckResourceAttr(resourceName, "cluster_admin_password", "T3JhY2xlVGVhbVVTQSExMjM="),
+				resource.TestCheckResourceAttrSet(resourceName, "cluster_public_key"),
+				resource.TestCheckResourceAttr(resourceName, "cluster_version", "ODH1"),
+				resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
+				resource.TestCheckResourceAttr(resourceName, "is_high_availability", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_secure", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_kafka_configured", "false"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.node_type"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.shape"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.subnet_id"),
+
+				func(s *terraform.State) (err error) {
+					resId, err = acctest.FromInstanceState(s, resourceName, "id")
+					if isEnableExportCompartment, _ := strconv.ParseBool(utils.GetEnvSettingWithDefault("enable_export_compartment", "true")); isEnableExportCompartment {
+						if errExport := resourcediscovery.TestExportCompartmentWithResourceName(&resId, &compartmentId, resourceName); errExport != nil {
+							return errExport
+						}
+					}
+					return err
+				},
+			),
+		},
 
 		// delete before next Create
 		{
@@ -316,7 +461,6 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				},
 			),
 		},
-
 		// verify Update to the compartment (the compartment will be switched back in the next step) and change shapes
 		{
 			Config: config + compartmentIdVariableStr + compartmentIdUVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + bootstrapScriptUrlVariableStr + BdsInstanceOdhResourceDependencies +
@@ -367,7 +511,6 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				},
 			),
 		},
-
 		// verify updates to updatable parameters, add a worker, update compute worker flex->regular, update util regular -> flex
 		{
 			Config: config + compartmentIdVariableStr + kmsKeyIdUVariableStr + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies +
@@ -404,7 +547,7 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				resource.TestCheckResourceAttr(resourceName, "util_node.0.shape", "VM.Standard.E4.Flex"),
 				resource.TestCheckResourceAttr(resourceName, "master_node.0.shape", "VM.Standard.E4.Flex"),
 				resource.TestCheckResourceAttr(resourceName, "compute_only_worker_node.0.shape", "VM.Standard2.4"),
-				resource.TestCheckResourceAttr(resourceName, "edge_node.0.shape", "VM.Standard2.4"),
+				resource.TestCheckResourceAttr(resourceName, "edge_node.0.shape", "VM.Standard.E4.Flex"),
 
 				func(s *terraform.State) (err error) {
 					resId2, err = acctest.FromInstanceState(s, resourceName, "id")
@@ -415,16 +558,65 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				},
 			),
 		},
+		// Add Master and Utility Node
+		{
+			Config: config + compartmentIdVariableStr + kmsKeyIdUVariableStr + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + subnetIdVariableStr + BdsInstanceOdhResourceDependencies +
+				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update, bdsInstanceOdhWithAddMasterUtilRepresentation),
+			Check: resource.ComposeAggregateTestCheckFunc(
+				resource.TestCheckResourceAttr(resourceName, "cluster_admin_password", "T3JhY2xlVGVhbVVTQSExMjM="),
+				resource.TestCheckResourceAttr(resourceName, "cluster_profile", "HADOOP"),
+				resource.TestCheckResourceAttrSet(resourceName, "cluster_public_key"),
+				resource.TestCheckResourceAttr(resourceName, "cluster_version", "ODH1"),
+				resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
+				resource.TestCheckResourceAttr(resourceName, "display_name", "displayName2"),
+				resource.TestCheckResourceAttr(resourceName, "freeform_tags.%", "1"),
+				resource.TestCheckResourceAttrSet(resourceName, "id"),
+				resource.TestCheckResourceAttrSet(resourceName, "is_cloud_sql_configured"),
+				resource.TestCheckResourceAttr(resourceName, "is_high_availability", "true"),
+				resource.TestCheckResourceAttr(resourceName, "is_secure", "true"),
+				resource.TestCheckResourceAttr(resourceName, "kerberos_realm_name", "BDSCLOUDSERVICE.ORACLE.COM"),
+				resource.TestCheckResourceAttr(resourceName, "kms_key_id", kmsKeyIdU),
+				resource.TestCheckResourceAttr(resourceName, "network_config.#", "1"),
+				resource.TestCheckResourceAttr(resourceName, "network_config.0.cidr_block", "111.112.0.0/16"),
+				resource.TestCheckResourceAttr(resourceName, "network_config.0.is_nat_gateway_required", "true"),
+				resource.TestCheckResourceAttr(resourceName, "nodes.#", "14"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.availability_domain"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.display_name"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.fault_domain"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.instance_id"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.node_type"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.shape"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.state"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.subnet_id"),
+				resource.TestCheckResourceAttrSet(resourceName, "nodes.0.time_created"),
+				resource.TestCheckResourceAttr(resourceName, "number_of_nodes", "14"),
+				resource.TestCheckResourceAttrSet(resourceName, "state"),
+				resource.TestCheckResourceAttr(resourceName, "util_node.0.shape", "VM.Standard.E4.Flex"),
+				resource.TestCheckResourceAttr(resourceName, "master_node.0.shape", "VM.Standard.E4.Flex"),
+				resource.TestCheckResourceAttr(resourceName, "compute_only_worker_node.0.shape", "VM.Standard.E4.Flex"),
+				resource.TestCheckResourceAttr(resourceName, "edge_node.0.shape", "VM.Standard.E4.Flex"),
+
+				func(s *terraform.State) (err error) {
+					resId2, err = acctest.FromInstanceState(s, resourceName, "id")
+					if resId != resId2 {
+						return fmt.Errorf("Resource recreated when it was supposed to be updated.")
+					}
+					return err
+				},
+			),
+		},
+
 		// verify datasource
 		{
 			Config: config +
 				acctest.GenerateDataSourceFromRepresentationMap("oci_bds_bds_instances", "test_bds_instances", acctest.Optional, acctest.Update, bdsInstanceOdhDataSourceRepresentation) +
 				compartmentIdVariableStr + kmsKeyIdUVariableStr + subnetIdVariableStr + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + BdsInstanceOdhResourceDependencies +
-				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update, bdsInstanceOdhWithRegularComputeAndFlexMasterUtilRepresentation),
+				acctest.GenerateResourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Optional, acctest.Update, bdsInstanceOdhWithAddMasterUtilRepresentation),
 			Check: resource.ComposeAggregateTestCheckFunc(
 				resource.TestCheckResourceAttr(datasourceName, "compartment_id", compartmentId),
 				resource.TestCheckResourceAttr(datasourceName, "display_name", "displayName2"),
 				resource.TestCheckResourceAttr(datasourceName, "state", "ACTIVE"),
+				resource.TestCheckResourceAttr(resourceName, "is_kafka_configured", "false"),
 				resource.TestCheckResourceAttr(datasourceName, "bds_instances.#", "1"),
 				resource.TestCheckResourceAttr(datasourceName, "bds_instances.0.cluster_profile", "HADOOP"),
 				resource.TestCheckResourceAttr(datasourceName, "bds_instances.0.cluster_version", "ODH1"),
@@ -436,6 +628,8 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				resource.TestCheckResourceAttr(datasourceName, "bds_instances.0.is_high_availability", "true"),
 				resource.TestCheckResourceAttr(datasourceName, "bds_instances.0.is_secure", "true"),
 				resource.TestCheckResourceAttrSet(datasourceName, "bds_instances.0.number_of_nodes"),
+				resource.TestCheckResourceAttrSet(datasourceName, "bds_instances.0.number_of_nodes_requiring_maintenance_reboot"),
+				resource.TestCheckResourceAttr(datasourceName, "bds_instances.0.state", "ACTIVE"),
 				resource.TestCheckResourceAttrSet(datasourceName, "bds_instances.0.state"),
 				resource.TestCheckResourceAttrSet(datasourceName, "bds_instances.0.time_created"),
 			),
@@ -444,7 +638,7 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 		{
 			Config: config +
 				acctest.GenerateDataSourceFromRepresentationMap("oci_bds_bds_instance", "test_bds_instance", acctest.Required, acctest.Create, bdsInstanceOdhSingularDataSourceRepresentation) +
-				compartmentIdVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + kmsKeyIdUVariableStr + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + BdsInstanceOdhWithRegularComputeWorkerResourceConfig,
+				compartmentIdVariableStr + kmsKeyIdVariableStr + subnetIdVariableStr + kmsKeyIdUVariableStr + bootstrapScriptUrlVariableStr + bootstrapScriptUrlUVariableStr + BdsInstanceOdhWithAddMasterUtilWorkerResourceConfig,
 			Check: resource.ComposeAggregateTestCheckFunc(
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "bds_instance_id"),
 
@@ -464,12 +658,13 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				resource.TestCheckResourceAttr(singularDatasourceName, "freeform_tags.%", "1"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "id"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "is_cloud_sql_configured"),
+				resource.TestCheckResourceAttr(resourceName, "is_kafka_configured", "false"),
 				resource.TestCheckResourceAttr(singularDatasourceName, "is_high_availability", "true"),
 				resource.TestCheckResourceAttr(singularDatasourceName, "is_secure", "true"),
 				resource.TestCheckResourceAttr(singularDatasourceName, "network_config.#", "1"),
 				resource.TestCheckResourceAttr(singularDatasourceName, "network_config.0.cidr_block", "111.112.0.0/16"),
 				resource.TestCheckResourceAttr(singularDatasourceName, "network_config.0.is_nat_gateway_required", "true"),
-				resource.TestCheckResourceAttr(singularDatasourceName, "nodes.#", "12"),
+				resource.TestCheckResourceAttr(singularDatasourceName, "nodes.#", "14"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "nodes.0.availability_domain"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "nodes.0.display_name"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "nodes.0.fault_domain"),
@@ -482,13 +677,16 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "nodes.0.state"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "nodes.0.time_created"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "number_of_nodes"),
+				resource.TestCheckResourceAttrSet(singularDatasourceName, "number_of_nodes_requiring_maintenance_reboot"),
+				resource.TestCheckResourceAttr(singularDatasourceName, "state", "ACTIVE"),
+				resource.TestCheckResourceAttrSet(singularDatasourceName, "number_of_nodes"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "state"),
 				resource.TestCheckResourceAttrSet(singularDatasourceName, "time_created"),
 			),
 		},
 		// verify resource import
 		{
-			Config:            config + BdsInstanceOdhWithRegularComputeRequiredOnlyResource,
+			Config:            config + BdsInstanceOdhWithAddMasterUtilRequiredOnlyResource,
 			ImportState:       true,
 			ImportStateVerify: true,
 			ImportStateVerifyIgnore: []string{
@@ -502,6 +700,7 @@ func TestResourceBdsOdhInstance(t *testing.T) {
 				"worker_node.0.shape_config",
 				"compute_only_worker_node.0.shape_config",
 				"edge_node.0.shape_config",
+				"kafka_broker_node.0.shape_config",
 			},
 			ResourceName: resourceName,
 		},